pyproject.toml
[project]
name = "hello-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.14",
    "langchain-openai>=0.3.0",
    "litellm>=1.58.0",
    "ruff>=0.9.1",
    "smolagents>=1.2.2",
]

[tool.ruff]
lint.extend-select = ["I"]
